+++ app/models/user.rb
@@ -1,4 +1,4 @@
 class User < ApplicationRecord
-  has_many :projects, foreign_key: "delivery_officer"
+  has_many :projects, foreign_key: "caseworker"
   has_many :notes
 end