+++ spec/features/users_can_update_projects_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.feature "Users can reach the edit project page" do
   let(:urn) { 12345 }
   let(:team_leader) { create(:user, :team_leader, email: "teamleader@education.gov.uk") }
-  let(:delivery_officer) { create(:user, email: "user1@education.gov.uk") }
+  let(:caseworker) { create(:user, email: "user1@education.gov.uk") }
   let(:unassigned_project) { create(:project, urn: urn) }
 
   before { mock_successful_api_responses(urn: urn, ukprn: 10061021) }
@@ -28,10 +28,10 @@ RSpec.feature "Users can reach the edit project page" do
   end
 
   context "the user is not a team leader" do
-    let(:user_1_project) { create(:project, urn: urn, delivery_officer: delivery_officer) }
+    let(:user_1_project) { create(:project, urn: urn, caseworker: caseworker) }
 
     before do
-      sign_in_with_user(delivery_officer)
+      sign_in_with_user(caseworker)
     end
 
     scenario "there is no edit link on the show project page" do
@@ -48,8 +48,8 @@ end
 
 RSpec.feature "Users can update a project" do
   let(:team_leader) { create(:user, :team_leader, email: "teamleader@education.gov.uk") }
-  let!(:delivery_officer) { create(:user, email: "user1@education.gov.uk") }
-  let!(:delivery_officer_2) { create(:user, email: "user2@education.gov.uk") }
+  let!(:caseworker) { create(:user, email: "user1@education.gov.uk") }
+  let!(:caseworker_2) { create(:user, email: "user2@education.gov.uk") }
 
   context "the user is a team leader" do
     before do
@@ -57,33 +57,33 @@ RSpec.feature "Users can update a project" do
       mock_successful_api_responses(urn: 12345, ukprn: 10061021)
     end
 
-    context "when the project has no delivery officer" do
+    context "when the project has no caseworker" do
       let(:unassigned_project) { create(:project) }
 
-      scenario "the delivery officer can be set" do
+      scenario "the caseworker can be set" do
         visit edit_project_path(unassigned_project)
-        select delivery_officer.email, from: "project-delivery-officer-id-field"
+        select caseworker.email, from: "project-caseworker-id-field"
         click_on "Continue"
-        expect(page).to have_content(delivery_officer.email)
+        expect(page).to have_content(caseworker.email)
       end
     end
 
-    context "when the project already has a delivery officer" do
-      let(:assigned_project) { create(:project, delivery_officer: delivery_officer) }
+    context "when the project already has a caseworker" do
+      let(:assigned_project) { create(:project, caseworker: caseworker) }
 
-      scenario "the delivery officer can be changed" do
-        delivery_officer_2
+      scenario "the caseworker can be changed" do
+        caseworker_2
         visit edit_project_path(assigned_project)
-        select delivery_officer_2.email, from: "project-delivery-officer-id-field"
+        select caseworker_2.email, from: "project-caseworker-id-field"
         click_on "Continue"
-        expect(page).to have_content(delivery_officer_2.email)
+        expect(page).to have_content(caseworker_2.email)
       end
 
-      scenario "the delivery officer can be unset" do
+      scenario "the caseworker can be unset" do
         visit edit_project_path(assigned_project)
-        select "", from: "project-delivery-officer-id-field"
+        select "", from: "project-caseworker-id-field"
         click_on "Continue"
-        expect(page).to have_content(I18n.t("project.summary.delivery_officer.unassigned"))
+        expect(page).to have_content(I18n.t("project.summary.caseworker.unassigned"))
       end
     end
   end