+++ spec/models/project_spec.rb
@@ -5,7 +5,7 @@ RSpec.describe Project, type: :model do
     it { is_expected.to have_db_column(:urn).of_type :integer }
     it { is_expected.to have_db_column(:trust_ukprn).of_type :integer }
     it { is_expected.to have_db_column(:target_completion_date).of_type :date }
-    it { is_expected.to have_db_column(:delivery_officer_id).of_type :uuid }
+    it { is_expected.to have_db_column(:caseworker_id).of_type :uuid }
     it { is_expected.to have_db_column(:team_leader_id).of_type :uuid }
   end
 
@@ -14,7 +14,7 @@ RSpec.describe Project, type: :model do
 
     it { is_expected.to have_many(:sections).dependent(:destroy) }
     it { is_expected.to have_many(:notes).dependent(:destroy) }
-    it { is_expected.to belong_to(:delivery_officer).required(false) }
+    it { is_expected.to belong_to(:caseworker).required(false) }
     it { is_expected.to belong_to(:team_leader).required(true) }
 
     describe "delete related entities" do