+++ CHANGELOG.md
@@ -38,6 +38,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   notes form.
 - Update the conversion workflow, Clear legal documents task list
 - The task list is now referred to as "Task list" not "Project task list"
+- Rename any references of "Delivery officer" to "Caseworker". Projects will
+  lose any currently assigned delivery officers.
 
 ## [Release 1][release-1]